+++ spicedb-python/examples/sync_read_relationships/test_sync_read_relationships.py
@@ -9,7 +9,7 @@ results.
 
 import pytest
 
-from conftest import SCHEMA
+from conftest import ENDPOINT, SCHEMA, TOKEN
 from spicedb import Filter, Relationship, full
 from spicedb.sync import SpiceDBClient
 from spicedb.types import Transaction
@@ -20,7 +20,7 @@ pytestmark = pytest.mark.integration
 
 def test_read_relationships_sync():
     with SpiceDBClient(
-        "localhost:50051", token="somerandomkeyhere", insecure=True
+        ENDPOINT, token=TOKEN, insecure=True
     ) as client:
         client.write_schema(SCHEMA)